% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dim.R
\name{get_dim}
\alias{get_dim}
\title{A Protein Sequence Function}
\usage{
get_dim(df)
}
\arguments{
\item{df}{data frame containing protein sequence information}
}
\value{
numeric length of the longest sequence
}
\description{
A Protein Sequence Function
}
\examples{
get_dim(df)
}

